requirements/base.txt
# Núcleo de la aplicación. Estas dependencias se usan en todos los entornos.

Django>=5.0,<5.2
djangorestframework>=3.15,<3.16

# Configuración por variables de entorno (12-factor)
django-environ>=0.11

# Driver PostgreSQL moderno (psycopg 3)
psycopg[binary]>=3.1

# Tareas asíncronas
celery>=5.4
redis>=5.0
django-redis>=5.4

# Servidor WSGI de producción
gunicorn>=22.0

# Cliente HTTP para los scanners
requests>=2.32

# Port scanning (wrapper sobre el binario nmap del sistema)
python-nmap>=0.7.1

# Parseo de certificados X.509 para el TLS scanner
cryptography>=42.0

# CORS (cuando llegue el frontend separado)
django-cors-headers>=4.4

# Hardening de auth
django-axes>=6.4          # Bloqueo por intentos fallidos
django-ratelimit>=4.1     # Rate limiting a nivel de vista

# PDF export
weasyprint>=60.0

# WHOIS / OSINT
python-whois>=0.9
